refactor: move user reads to user/api/userReads
